+++ grafana-plugin/src/containers/IntegrationTemplate/IntegrationTemplate.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useState, useEffect } from 'react';
 
-import { Button, HorizontalGroup, Drawer, VerticalGroup, Icon } from '@grafana/ui';
+import { Button, HorizontalGroup, Drawer, VerticalGroup } from '@grafana/ui';
 import cn from 'classnames/bind';
 import { debounce } from 'lodash-es';
 import { observer } from 'mobx-react';
@@ -12,10 +12,8 @@ import {
   slackMessageTemplateCheatSheet,
   genericTemplateCheatSheet,
 } from 'components/CheatSheet/CheatSheet.config';
-import Block from 'components/GBlock/Block';
 import MonacoEditor from 'components/MonacoEditor/MonacoEditor';
 import Text from 'components/Text/Text';
-import TemplatePreview from 'containers/TemplatePreview/TemplatePreview';
 import TemplatesAlertGroupsList, { TEMPLATE_PAGE } from 'containers/TemplatesAlertGroupsList/TemplatesAlertGroupsList';
 import { WithPermissionControlTooltip } from 'containers/WithPermissionControl/WithPermissionControlTooltip';
 import { AlertReceiveChannel } from 'models/alert_receive_channel/alert_receive_channel.types';
@@ -28,6 +26,7 @@ import LocationHelper from 'utils/LocationHelper';
 import { UserActions } from 'utils/authorization';
 
 import styles from './IntegrationTemplate.module.scss';
+import TemplateResult from 'containers/TemplateResult/TemplateResult';
 
 const cx = cn.bind(styles);
 
@@ -44,9 +43,6 @@ interface IntegrationTemplateProps {
 
 const IntegrationTemplate = observer((props: IntegrationTemplateProps) => {
   const { id, onHide, template, onUpdateTemplates, onUpdateRoute, templateBody, channelFilterId, templates } = props;
-
-  console.log('IntegrationTemplate');
-  console.log(props);
 
   const [isCheatSheetVisible, setIsCheatSheetVisible] = useState<boolean>(false);
   const [chatOpsPermalink, setChatOpsPermalink] = useState(undefined);
@@ -199,7 +195,7 @@ const IntegrationTemplate = observer((props: IntegrationTemplateProps) => {
             onLoadAlertGroupsList={onLoadAlertGroupsList}
           />
           {renderCheatSheet()}
-          <Result
+          <TemplateResult
             alertReceiveChannelId={id}
             template={template}
             templateBody={changedTemplateBody}
@@ -251,87 +247,5 @@ const IntegrationTemplate = observer((props: IntegrationTemplateProps) => {
     );
   }
 });
-
-interface ResultProps {
-  alertReceiveChannelId: AlertReceiveChannel['id'];
-  // templateName: string;
-  templateBody: string;
-  template: TemplateForEdit;
-  isAlertGroupExisting?: boolean;
-  chatOpsPermalink?: string;
-  payload?: JSON;
-  error?: string;
-  onSaveAndFollowLink?: (link: string) => void;
-  templateIsRoute?: boolean;
-}
-
-const Result = (props: ResultProps) => {
-  const {
-    alertReceiveChannelId,
-    template,
-    templateBody,
-    chatOpsPermalink,
-    payload,
-    error,
-    isAlertGroupExisting,
-    onSaveAndFollowLink,
-  } = props;
-
-  return (
-    <div className={cx('template-block-result')}>
-      <div className={cx('template-block-title')}>
-        <HorizontalGroup justify="space-between">
-          <Text>Result</Text>
-        </HorizontalGroup>
-      </div>
-      <div className={cx('result')}>
-        {payload || error ? (
-          <VerticalGroup spacing="lg">
-            {error ? (
-              <Block bordered fullWidth withBackground>
-                <Text>{error}</Text>
-              </Block>
-            ) : (
-              <Block bordered fullWidth withBackground>
-                <TemplatePreview
-                  key={template.name}
-                  templateName={template.name}
-                  templateBody={templateBody}
-                  templateType={template.type}
-                  templateIsRoute={template.isRoute}
-                  alertReceiveChannelId={alertReceiveChannelId}
-                  payload={payload}
-                />
-              </Block>
-            )}
-
-            {template?.additionalData?.additionalDescription && (
-              <Text type="secondary">{template?.additionalData.additionalDescription}</Text>
-            )}
-
-            {template?.additionalData?.chatOpsName && isAlertGroupExisting && (
-              <VerticalGroup>
-                <Button onClick={() => onSaveAndFollowLink(chatOpsPermalink)}>
-                  <HorizontalGroup spacing="xs" align="center">
-                    Save and open Alert Group in {template.additionalData.chatOpsDisplayName}{' '}
-                    <Icon name="external-link-alt" />
-                  </HorizontalGroup>
-                </Button>
-
-                {template.additionalData.data && <Text type="secondary">{template.additionalData.data}</Text>}
-              </VerticalGroup>
-            )}
-          </VerticalGroup>
-        ) : (
-          <div>
-            <Block bordered fullWidth className={cx('block-style')} withBackground>
-              <Text>← Select alert group or "Use custom payload"</Text>
-            </Block>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
 
 export default IntegrationTemplate;
